Replace existing resource with new one.
